Fix interpolation for log group vairable
