+++ public/app/plugins/datasource/cloudwatch/language/cloudwatch-logs-sql/completion/CompletionItemProvider.ts
@@ -7,6 +7,7 @@ import { CompletionItemProvider } from '../../monarch/CompletionItemProvider';
 import { LinkedToken } from '../../monarch/LinkedToken';
 import { TRIGGER_SUGGEST } from '../../monarch/commands';
 import { SuggestionKind, CompletionItemPriority, StatementPosition } from '../../monarch/types';
+import { fetchLogGroupFields } from '../../utils';
 import {
   ASC,
   BY,
@@ -268,7 +269,7 @@ export class LogsSQLCompletionItemProvider extends CompletionItemProvider {
           break;
 
         case SuggestionKind.Field:
-          const fields = await this.fetchFields(this.queryContext.logGroups || [], this.queryContext.region);
+          const fields = await fetchLogGroupFields(this.queryContext.logGroups || [], this.queryContext.region, this.templateSrv, this.resources);
           fields.forEach((field) => {
             if (field !== '') {
               addSuggestion(field, {
@@ -295,20 +296,4 @@ export class LogsSQLCompletionItemProvider extends CompletionItemProvider {
 
     return suggestions;
   }
-
-  private fetchFields = async (logGroups: LogGroup[], region: string): Promise<string[]> => {
-    if (logGroups.length === 0) {
-      return [];
-    }
-
-    const results = await Promise.all(
-      logGroups.map((logGroup) =>
-        this.resources
-          .getLogGroupFields({ logGroupName: logGroup.name, arn: logGroup.arn, region })
-          .then((fields) => fields.filter((f) => f).map((f) => f.value.name ?? ''))
-      )
-    );
-    // Deduplicate fields
-    return [...new Set(results.flat())];
-  };
 }
