+++ public/app/plugins/datasource/cloudwatch/language/logs/completion/CompletionItemProvider.ts
@@ -7,6 +7,7 @@ import { CompletionItemProvider } from '../../monarch/CompletionItemProvider';
 import { LinkedToken } from '../../monarch/LinkedToken';
 import { TRIGGER_SUGGEST } from '../../monarch/commands';
 import { CompletionItem, CompletionItemPriority, StatementPosition, SuggestionKind } from '../../monarch/types';
+import { fetchLogGroupFields } from '../../utils';
 import { LOGS_COMMANDS, LOGS_FUNCTION_OPERATORS, SORT_DIRECTION_KEYWORDS } from '../language';
 
 import { getStatementPosition } from './statementPosition';
@@ -86,7 +87,7 @@ export class LogsCompletionItemProvider extends CompletionItemProvider {
           });
 
           if (this.queryContext.logGroups && this.queryContext.logGroups.length > 0) {
-            let fields = await this.fetchFields(this.queryContext.logGroups, this.queryContext.region);
+            let fields = await fetchLogGroupFields(this.queryContext.logGroups, this.queryContext.region, this.templateSrv, this.resources);
             fields.push('@log');
             fields.forEach((field) => {
               if (field !== '') {
@@ -133,16 +134,4 @@ export class LogsCompletionItemProvider extends CompletionItemProvider {
 
     return suggestions;
   }
-
-  private fetchFields = async (logGroups: LogGroup[], region: string): Promise<string[]> => {
-    const results = await Promise.all(
-      logGroups.map((logGroup) =>
-        this.resources
-          .getLogGroupFields({ logGroupName: logGroup.name, arn: logGroup.arn, region })
-          .then((fields) => fields.filter((f) => f).map((f) => f.value.name ?? ''))
-      )
-    );
-    // Deduplicate fields
-    return [...new Set(results.flat())];
-  };
 }
