+++ public/app/plugins/datasource/cloudwatch/resources/ResourcesAPI.ts
@@ -12,7 +12,6 @@ import {
   ResourceResponse,
   DescribeLogGroupsRequest,
   LogGroupResponse,
-  GetLogGroupFieldsRequest,
   GetMetricsRequest,
   GetDimensionKeysRequest,
   GetDimensionValuesRequest,
@@ -79,15 +78,13 @@ export class ResourcesAPI extends CloudWatchRequest {
     });
   }
 
-  getLogGroupFields({
-    region,
-    arn,
-    logGroupName,
-  }: GetLogGroupFieldsRequest): Promise<Array<ResourceResponse<LogGroupField>>> {
+  getLogGroupFields(
+    region: string,
+    logGroupName: string
+  ): Promise<Array<ResourceResponse<LogGroupField>>> {
     return this.memoizedGetRequest<Array<ResourceResponse<LogGroupField>>>('log-group-fields', {
       region: this.templateSrv.replace(this.getActualRegion(region)),
-      logGroupName: this.templateSrv.replace(logGroupName, {}),
-      logGroupArn: this.templateSrv.replace(arn),
+      logGroupName: logGroupName,
     });
   }
